online_inference/requirements.txt
anyio==3.6.2
attrs==22.1.0
boto3==1.23.10
botocore==1.26.10
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
fastapi==0.87.0
h11==0.14.0
httpcore==0.16.1
httpx==0.23.1
idna==3.4
iniconfig==1.1.1
jmespath==1.0.1
joblib==1.2.0
numpy==1.23.5
packaging==21.3
pandas==1.5.1
pluggy==1.0.0
pydantic==1.10.2
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
rfc3986==1.5.0
s3transfer==0.5.0
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
sniffio==1.3.0
starlette==0.21.0
threadpoolctl==3.1.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.12
uvicorn==0.19.0
